+++ app/(user)/checkout/page.tsx
@@ -1,21 +1,17 @@
 "use client";
 import { useState, useRef, useEffect } from "react";
 import Link from "next/link";
-import { Search, User, Heart, ShoppingBag, ChevronDown } from "react-feather";
 import Footer from "../../components/Footer";
 import Header from "../../components/Header";
 import { useCart, CartItem } from "../../context/CartContext";
 import { getValidImageUrl, handleImageError } from "@/lib/imageUtils";
-import { useRouter } from "next/navigation";
+import Success from "@/app/components/checkout/Success";
+import Cancel from "@/app/components/checkout/Cancel";
 
 function Checkout() {
   const [shipping, setShipping] = useState("ship");
-  const [searchQuery, setSearchQuery] = useState("");
-  const [openDropdown, setOpenDropdown] = useState<string | null>(null);
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
   const [isClient, setIsClient] = useState(false);
   const { cart, getCartTotal, isLoading, clearCart } = useCart();
-  const router = useRouter();
   const cartClearedRef = useRef(false);
 
   const [formData, setFormData] = useState({
@@ -85,8 +81,8 @@ function Checkout() {
       ...formData,
       cart: cart,
       subtotal: getCartTotal(),
-      shippingCost: shipping === "ship" ? 650 : 0,
-      total: shipping === "ship" ? getCartTotal() + 650 : getCartTotal(),
+      shippingCost: shipping === "ship" ? 10 : 0,
+      total: shipping === "ship" ? getCartTotal() + 10 : getCartTotal(),
     };
 
     console.log("Submitting checkout data:", checkoutData);
@@ -135,19 +131,6 @@ function Checkout() {
     setIsClient(true);
   }, []);
 
-  const handleMouseEnter = (category: string) => {
-    if (timeoutRef.current !== null) {
-      clearTimeout(timeoutRef.current);
-    }
-    setOpenDropdown(category);
-  };
-
-  const handleMouseLeave = () => {
-    timeoutRef.current = setTimeout(() => {
-      setOpenDropdown(null);
-    }, 200);
-  };
-
   const getDisplayColorName = (color?: string): string => {
     if (!color) return "Default";
 
@@ -172,34 +155,18 @@ function Checkout() {
     );
   }
 
-  // Checking URL parameters directly in the render function for success message
-  const isSuccess =
-    typeof window !== "undefined" &&
-    new URLSearchParams(window.location.search).get("success") === "1";
-
-  if (isSuccess) {
+  // Checking URL parameters for success or cancel
+  if (typeof window !== "undefined" && new URLSearchParams(window.location.search).get("success") === "1") {
     const orderNumber = new URLSearchParams(window.location.search).get(
       "order"
     );
     return (
-      <div className="min-h-screen flex flex-col bg-gray-50">
-        <Header />
-        <div className="m-3 flex items-center justify-center bg-gray-50">
-          <div className="p-6 bg-white rounded-lg shadow-md text-center">
-            <h1 className="text-2xl font-bold text-green-600 mb-4">Success!</h1>
-            <p className="text-lg text-gray-700">
-              Payment successful! Your order number is {orderNumber}.
-            </p>
-            <button
-              onClick={() => router.push("/")}
-              className="mt-6 px-4 py-2 bg-black text-white rounded-md hover:bg-gray-900 transition"
-            >
-              Go to Homepage
-            </button>
-          </div>
-        </div>
-        <Footer />
-      </div>
+      <Success orderNumber={orderNumber || "N/A"} />
+    );
+  }
+  else if (typeof window !== "undefined" && new URLSearchParams(window.location.search).get("canceled") === "1"){
+    return (
+      <Cancel />
     );
   }
 
@@ -297,11 +264,11 @@ function Checkout() {
                   <>
                     <div className="flex justify-between py-1">
                       <span className="text-gray-600">Shipping</span>
-                      <span className="font-medium">$650.00</span>
+                      <span className="font-medium">$10.00</span>
                     </div>
                     <div className="flex justify-between py-3 text-lg font-semibold border-t border-gray-200 mt-2">
                       <span>Total</span>
-                      <span>${(getCartTotal() + 650).toFixed(2)}</span>
+                      <span>${(getCartTotal() + 10).toFixed(2)}</span>
                     </div>
                   </>
                 ) : (
@@ -639,7 +606,7 @@ function Checkout() {
                         />
                         <span>Standard Shipping</span>
                       </div>
-                      <span className="font-medium">$650.00</span>
+                      <span className="font-medium">$10.00</span>
                     </label>
                   ) : (
                     <div className="p-4 border border-gray-300 rounded-lg bg-gray-50">
@@ -674,25 +641,6 @@ function Checkout() {
                     </div>
                   )}
                 </section>
-
-                <input
-                  type="hidden"
-                  name="cartItems"
-                  value={JSON.stringify(cart)}
-                />
-                <input
-                  type="hidden"
-                  name="subtotal"
-                  value={getCartTotal().toString()}
-                />
-                <input
-                  type="hidden"
-                  name="total"
-                  value={(shipping === "ship"
-                    ? getCartTotal() + 650
-                    : getCartTotal()
-                  ).toString()}
-                />
 
                 <button
                   type="submit"
